Rename `oldtime` module to `duration`
